# backend/api/openapi.yml
openapi: 3.0.3
info:
  title: Where Did My Money Go API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: User registered
        "409":
          description: Email already exists
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "204":
          description: Logged in
          headers:
            Set-Cookie:
              schema:
                type: string
  /auth/logout:
    post:
      responses:
        "204":
          description: Logged out
  /expenses:
    get:
      summary: List expenses for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpense"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /summary:
    get:
      summary: Category summary for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Summary per category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

components:
  schemas:
    Expense:
      type: object
      required: [id, category, amount, date]
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date-time
    CreateExpense:
      type: object
      required: [category, amount, date]
      properties:
        category:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date-time
    CategorySummary:
      type: object
      required: [category, total]
      properties:
        category:
          type: string
        total:
          type: integer